+++ NEWS.md
@@ -5,18 +5,18 @@
 
 # specklestar v0.0.1.8 (2017-03-06)
 
-* Added par_angle() and par_angle_manual() R functions for
+* Added par_angle_BTA() and par_angle_log() R functions for
 calculation of parallactic angles for BTA location
 * Added database constrution from folder with
 BTA observational journals
 * Added speckle_rho() R function for convertation pixels to arc seconds
 * Added reading all types of binary files in speckle_frame() function
 * Added annulus() R function to get points inside of annulus on matrix
+* Added calculation of detection limits
 
 # Roadmap of specklestar package
 
 - Seeing calculation
-- Calculation of detection limits
 - Remove log_date from par_angle.R
 - Calculation of binary star orbit
 - Function to find position of secondary maximum on ACF